+++ src/RouteManager.php
@@ -17,37 +17,24 @@ class RouteManager
 
 	private $errorRoute;
 
-	private $simpleRoute;
-
-	private $dynamicRoute;
-
-	public function __construct(
-		RouteContainer $plugRoute,
-		Request $request,
-		SimpleRoute $simpleRoute,
-		DynamicRoute $dynamicRoute
-	)
+	public function __construct(RouteContainer $plugRoute, Request $request)
 	{
 		$this->request 		= $request;
 		$routes 			= $plugRoute->getRoutes();
 		$this->callback     = new Callback($this->request);
 		$this->routes       = $routes[$this->request->method()];
-		$this->simpleRoute	= $simpleRoute;
-		$this->dynamicRoute	= $dynamicRoute;
 		$this->errorRoute	= $plugRoute->getErrorRouteNotFound();
 	}
 
 	public function run(array $dependencies = [])
 	{
-		$this->dynamicRoute->next($this->simpleRoute);
 		$url = $this->request->getUrl();
+        $analyzer = new RouteAnalyzer();
 
 		foreach ($this->routes as $route) {
-			$routerObject = $this->dynamicRoute->handler($route->getRoute(), $url);
-			$routeHandled = $routerObject->getRoute();
-
+			$routeHandled = $analyzer->getRoute($route->getRoute(), $url);
 			if (ValidateHelper::isEqual($routeHandled, $url)) {
-				$this->setParameters($routerObject->getParameters());
+				$this->setParameters($analyzer->getParameters());
 
 				return $this->callback->handlerCallback($route, $dependencies);
 			}
@@ -62,9 +49,9 @@ class RouteManager
 			return $this->callback->handlerCallback($this->errorRoute);
 		}
 
-		$response = new Response();
-		$response->setStatusCode(404)->response();
-		return Error::throwException("The route could not be found.");
+        $response = new Response();
+        $response->setStatusCode(404)->response();
+		echo "The route could not be found.";
 	}
 
 	private function setParameters($parameters)