+++ doc/changelog.md
@@ -1,5 +1,13 @@
 # Changelog
 
+### v4.1
+* Now, It's possible get parameter without use Request class
+* Refactored RouteAnalyzer
+* Deleted DynamicRoute and SimpleRoute classes
+* Created RouteFactory class
+* Refactored RouteManager, PlugRoute and MatchHelper classes
+* Refactored tests
+
 ### v4.0
 * Refactored Callback class
 * Refactored PlugRouteMiddleware interface